Move methods to non-detector specific UI.
