+++ chemclipse/plugins/org.eclipse.chemclipse.chromatogram.vsd.filter.ui/src/org/eclipse/chemclipse/chromatogram/vsd/filter/ui/services/WavenumberSignalsAnnotationService.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2021, 2024 Lablicate GmbH.
+ * Copyright (c) 2021, 2025 Lablicate GmbH.
  *
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
@@ -15,7 +15,7 @@ import java.util.concurrent.atomic.AtomicReference;
 
 import org.eclipse.chemclipse.chromatogram.vsd.filter.service.WavenumberSignalsSerializationService;
 import org.eclipse.chemclipse.chromatogram.vsd.filter.ui.swt.WavenumberSignalsEditor;
-import org.eclipse.chemclipse.ux.extension.xxd.ui.methods.IAnnotationWidgetService;
+import org.eclipse.chemclipse.ux.extension.ui.methods.IAnnotationWidgetService;
 import org.eclipse.swt.SWT;
 import org.eclipse.swt.layout.GridData;
 import org.eclipse.swt.widgets.Composite;
