+++ chemclipse/plugins/org.eclipse.chemclipse.ux.extension.ui/src/org/eclipse/chemclipse/ux/extension/ui/l10n/Messages.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2023, 2024 Lablicate GmbH.
+ * Copyright (c) 2023, 2025 Lablicate GmbH.
  *
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
@@ -17,7 +17,7 @@ import org.eclipse.osgi.util.NLS;
 public class Messages extends NLS {
 
 	private static final String BUNDLE_NAME = "org.eclipse.chemclipse.ux.extension.ui.l10n.messages"; //$NON-NLS-1$ //
-	//
+
 	public static String drives;
 	public static String home;
 	public static String openAllContainedMeasurements;
@@ -32,7 +32,33 @@ public class Messages extends NLS {
 	public static String workspace;
 	public static String setAsMethodDirectory;
 	public static String setAsActiveMethod;
-	//
+	public static String literatureReferences;
+	public static String literatureReference;
+	public static String noLinkIsSupplierYet;
+	public static String openInExternalBrowser;
+	public static String chooseLocation;
+	public static String processorOffersNoOptions;
+	public static String rememberDecision;
+	public static String resetDefaults;
+	public static String settings;
+	public static String collapseAllProcessorItems;
+	public static String displaySettings;
+	public static String yes;
+	public static String no;
+	public static String editProcessorOptions;
+	public static String editStoredDataForSelectedProcessor;
+	public static String expandAllProcessorItems;
+	public static String managePreferences;
+	public static String manageProcessorOptions;
+	public static String options;
+	public static String systemDefault;
+	public static String processorOptionsBelowSelectToManageRemoveState;
+	public static String removeStoredDataResetDefaultsForSelectedProcessor;
+	public static String removeAllStoredDataResetDefaults;
+	public static String resetSettingsForAllProcessors;
+	public static String resetSettingsForSelectedProcessors;
+	public static String selectOptionsForProcessorName;
+
 	static {
 		NLS.initializeMessages(BUNDLE_NAME, Messages.class);
 	}
