+++ chemclipse/plugins/org.eclipse.chemclipse.ux.extension.ui/src/org/eclipse/chemclipse/ux/extension/ui/l10n/messages.properties
@@ -12,3 +12,28 @@ userLocation=User Location
 workspace=Workspace
 setAsMethodDirectory=Set as method directory
 setAsActiveMethod=Set as active method
+editStoredDataForSelectedProcessor=Edit the stored data for the selected processor.
+removeStoredDataResetDefaultsForSelectedProcessor=Removes the stored data for the selected processor and resets it to the defaults.
+removeAllStoredDataResetDefaults=Removes all stored data and resets to the defaults.
+resetSettingsForAllProcessors=Would you like to reset the settings for all processors?
+expandAllProcessorItems=Expand all processor items.
+collapseAllProcessorItems=Collapse all processor items.
+displaySettings=Display Settings
+yes=yes
+no=no
+options=Options
+systemDefault=System Default
+resetSettingsForSelectedProcessors=Would you like to reset the settings for the selected processors?
+rememberDecision=Remember my decision and don't ask again.
+resetDefaults=Restore Defaults
+processorOffersNoOptions=This processor offers no options.
+editProcessorOptions=Edit Processor Options
+selectOptionsForProcessorName=Select the options for {0}
+manageProcessorOptions=Manage Processor Options
+managePreferences=Manage Preferences
+processorOptionsBelowSelectToManageRemoveState=Below you find all currently stored processor Options, select one to manage or remove the stored state
+chooseLocation=Please choose a location ...
+literatureReferences=Literature References
+literatureReference=Literature Reference
+noLinkIsSupplierYet=No link is supplied yet.
+openInExternalBrowser=Open link in an external browser.
