+++ chemclipse/plugins/org.eclipse.chemclipse.ux.extension.ui/src/org/eclipse/chemclipse/ux/extension/ui/l10n/messages_de.properties
@@ -12,3 +12,28 @@ userLocation=Benutzerverzeichnis
 workspace=Arbeitsbereich
 setAsMethodDirectory=Methodenverzeichnis auswählen
 setAsActiveMethod=Methode aktivieren
+editStoredDataForSelectedProcessor=Bearbeitet die gespeicherten Daten für den ausgewählten Prozessor.
+removeStoredDataResetDefaultsForSelectedProcessor=Entfernt die gespeicherten Daten für den ausgewählten Prozessor und setzt sie auf die Standardwerte zurück.
+removeAllStoredDataResetDefaults=Entfernt alle gespeicherten Daten und setzt auf die Standardeinstellungen zurück.
+resetSettingsForAllProcessors=Sollen die Einstellungen für alle Prozessoren zurückgesetzt werden?
+expandAllProcessorItems=Erweitert alle Prozessorelemente.
+collapseAllProcessorItems=Blenden Sie alle Prozessorelemente aus.
+displaySettings=Anzeigeeinstellungen
+yes=ja
+no=nein
+options=Optionen
+systemDefault=Systemstandard
+resetSettingsForSelectedProcessors=Möchten Sie die Einstellungen für die ausgewählten Prozessoren zurücksetzen?
+rememberDecision=Einstellungen speichern und nicht erneut fragen.
+resetDefaults=Standardwerte wiederherstellen.
+processorOffersNoOptions=Dieser Prozessor bietet keine Optionen.
+editProcessorOptions=Bearbeite Prozessoroptionen
+selectOptionsForProcessorName=Wählen Sie die Optionen für {0}
+manageProcessorOptions=Prozessoroptionen verwalten
+managePreferences=Einstellungen verwalten
+processorOptionsBelowSelectToManageRemoveState=Unten finden Sie alle derzeit gespeicherten Prozessoroptionen, wählen Sie eine aus, um den gespeicherten Zustand zu verwalten oder zu entfernen
+chooseLocation=Bitte wählen Sie einen Ort ...
+literatureReferences=Literaturstellen
+literatureReference=Literaturstelle
+noLinkIsSupplierYet=Kein Link vorhanden.
+openInExternalBrowser=Link in externem Browser öffnen.
