+++ chemclipse/plugins/org.eclipse.chemclipse.ux.extension.ui/src/org/eclipse/chemclipse/ux/extension/ui/methods/IAnnotationWidgetService.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2021 Lablicate GmbH.
+ * Copyright (c) 2021, 2025 Lablicate GmbH.
  *
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
@@ -9,11 +9,15 @@
  * Contributors:
  * Philip Wenig - initial API and implementation
  *******************************************************************************/
-package org.eclipse.chemclipse.ux.extension.xxd.ui.methods;
+package org.eclipse.chemclipse.ux.extension.ui.methods;
 
-import org.eclipse.swt.widgets.Listener;
+import org.eclipse.chemclipse.support.settings.serialization.ISerializationService;
+import org.eclipse.swt.widgets.Composite;
+import org.eclipse.swt.widgets.Control;
 
-public interface IChangeListener {
+public interface IAnnotationWidgetService extends ISerializationService {
 
-	void addChangeListener(Listener listener);
+	Control createWidget(Composite parent, String description, Object currentSelection);
+
+	Object getValue(Object currentSelection);
 }
