+++ chemclipse/plugins/org.eclipse.chemclipse.ux.extension.xxd.ui/OSGI-INF/org.eclipse.chemclipse.ux.extension.xxd.ui.services.TimeRangesAnnotationService.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <scr:component xmlns:scr="http://www.osgi.org/xmlns/scr/v1.1.0" configuration-policy="optional" name="org.eclipse.chemclipse.ux.extension.xxd.ui.services.TimeRangesAnnotationService">
    <service>
-      <provide interface="org.eclipse.chemclipse.ux.extension.xxd.ui.methods.IAnnotationWidgetService"/>
+      <provide interface="org.eclipse.chemclipse.ux.extension.ui.methods.IAnnotationWidgetService"/>
    </service>
    <implementation class="org.eclipse.chemclipse.ux.extension.xxd.ui.services.TimeRangesAnnotationService"/>
 </scr:component>
