+++ chemclipse/plugins/org.eclipse.chemclipse.ux.extension.xxd.ui/src/org/eclipse/chemclipse/ux/extension/xxd/ui/l10n/messages.properties
@@ -254,28 +254,7 @@ resumeFollowingEntryDefaultComplete=Resume processing at the following entry. By
 completeMethod=Complete Method
 rememberDecisionDontShowAgain=Remember my decision and don't show the dialog again.
 revertDecisionInSettings=You can revert this decision in the settings.
-editStoredDataForSelectedProcessor=Edit the stored data for the selected processor.
-removeStoredDataResetDefaultsForSelectedProcessor=Removes the stored data for the selected processor and resets it to the defaults.
-removeAllStoredDataResetDefaults=Removes all stored data and resets to the defaults.
-resetSettingsForAllProcessors=Would you like to reset the settings for all processors?
-expandAllProcessorItems=Expand all processor items.
-collapseAllProcessorItems=Collapse all processor items.
-displaySettings=Display Settings
-yes=yes
-no=no
-options=Options
-systemDefault=System Default
-resetSettingsForSelectedProcessors=Would you like to reset the settings for the selected processors?
-rememberDecision=Remember my decision and don't ask again.
-resetDefaults=Restore Defaults
 resetPlate=Reset the plate
-processorOffersNoOptions=This processor offers no options.
-editProcessorOptions=Edit Processor Options
-selectOptionsForProcessorName=Select the options for {0}
-manageProcessorOptions=Manage Processor Options
-managePreferences=Manage Preferences
-processorOptionsBelowSelectToManageRemoveState=Below you find all currently stored processor Options, select one to manage or remove the stored state
-chooseLocation=Please choose a location ...
 text=Text
 namedTrace=Named Trace
 unknown=Unknown
@@ -330,10 +309,6 @@ exportFailed=Export failed
 fluorescence=Fluorescence
 pcrExport=PCR Export
 plateChart=Plate Chart
-literatureReferences=Literature References
-literatureReference=Literature Reference
-noLinkIsSupplierYet=No link is supplied yet.
-openInExternalBrowser=Open link in an external browser.
 transmittance=Transmittance
 absorbance=Absorbance
 wavenumber=Wavenumber [1/cm]
