+++ chemclipse/plugins/org.eclipse.chemclipse.ux.extension.xxd.ui/src/org/eclipse/chemclipse/ux/extension/xxd/ui/l10n/messages_de.properties
@@ -251,28 +251,7 @@ resumeFollowingEntryDefaultComplete=Prozessierung am folgenden Eintrag fortsetze
 completeMethod=vollständige Methode
 rememberDecisionDontShowAgain=Entscheidung merken und diesen Dialog nicht mehr anzeigen.
 revertDecisionInSettings=Die Entscheidung kann in den Einstellung rückgängig gemacht werden.
-editStoredDataForSelectedProcessor=Bearbeitet die gespeicherten Daten für den ausgewählten Prozessor.
-removeStoredDataResetDefaultsForSelectedProcessor=Entfernt die gespeicherten Daten für den ausgewählten Prozessor und setzt sie auf die Standardwerte zurück.
-removeAllStoredDataResetDefaults=Entfernt alle gespeicherten Daten und setzt auf die Standardeinstellungen zurück.
-resetSettingsForAllProcessors=Sollen die Einstellungen für alle Prozessoren zurückgesetzt werden?
-expandAllProcessorItems=Erweitert alle Prozessorelemente.
-collapseAllProcessorItems=Blenden Sie alle Prozessorelemente aus.
-displaySettings=Anzeigeeinstellungen
-yes=ja
-no=nein
-options=Optionen
-systemDefault=Systemstandard
-resetSettingsForSelectedProcessors=Möchten Sie die Einstellungen für die ausgewählten Prozessoren zurücksetzen?
-rememberDecision=Einstellungen speichern und nicht erneut fragen.
-resetDefaults=Standardwerte wiederherstellen.
 resetPlate=Platte zurücksetzen.
-processorOffersNoOptions=Dieser Prozessor bietet keine Optionen.
-editProcessorOptions=Bearbeite Prozessoroptionen
-selectOptionsForProcessorName=Wählen Sie die Optionen für {0}
-manageProcessorOptions=Prozessoroptionen verwalten
-managePreferences=Einstellungen verwalten
-processorOptionsBelowSelectToManageRemoveState=Unten finden Sie alle derzeit gespeicherten Prozessoroptionen, wählen Sie eine aus, um den gespeicherten Zustand zu verwalten oder zu entfernen
-chooseLocation=Bitte wählen Sie einen Ort ...
 text=Text
 namedTrace=Benannte Spur
 unknown=Unbekannt
@@ -327,10 +306,6 @@ exportFailed=Export fehlgeschlagen
 fluorescence=Fluoreszenz
 pcrExport=PCR Export
 plateChart=Plattendiagramm
-literatureReferences=Literaturreferenzen
-literatureReference=Literaturreferenz
-noLinkIsSupplierYet=Kein Link vorhanden.
-openInExternalBrowser=Link in externem Browser öffnen.
 transmittance=Transmission
 absorbance=Absorbtion
 wavenumber=Wellenzahl [1/cm]
