+++ chemclipse/plugins/org.eclipse.chemclipse.ux.extension.xxd.ui/src/org/eclipse/chemclipse/ux/extension/xxd/ui/services/ColumnMappingAnnotationService.java
@@ -12,7 +12,7 @@
 package org.eclipse.chemclipse.ux.extension.xxd.ui.services;
 
 import org.eclipse.chemclipse.model.service.ColumnMappingSerializationService;
-import org.eclipse.chemclipse.ux.extension.xxd.ui.methods.IAnnotationWidgetService;
+import org.eclipse.chemclipse.ux.extension.ui.methods.IAnnotationWidgetService;
 import org.eclipse.chemclipse.ux.extension.xxd.ui.swt.editors.ColumnMappingEditorUI;
 import org.eclipse.swt.SWT;
 import org.eclipse.swt.layout.GridData;
