+++ chemclipse/plugins/org.eclipse.chemclipse.xxd.classification.ui/src/org/eclipse/chemclipse/xxd/classification/ui/swt/ClassificationDictionaryEditor.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2022, 2024 Lablicate GmbH.
+ * Copyright (c) 2022, 2025 Lablicate GmbH.
  * 
  * All rights reserved.
  * This program and the accompanying materials are made available under the
@@ -22,8 +22,8 @@ import org.eclipse.chemclipse.support.ui.swt.ITableSettings;
 import org.eclipse.chemclipse.support.updates.IUpdateListener;
 import org.eclipse.chemclipse.swt.ui.components.ISearchListener;
 import org.eclipse.chemclipse.swt.ui.components.SearchSupportUI;
+import org.eclipse.chemclipse.ux.extension.ui.methods.IChangeListener;
 import org.eclipse.chemclipse.ux.extension.ui.swt.IExtendedPartUI;
-import org.eclipse.chemclipse.ux.extension.xxd.ui.methods.IChangeListener;
 import org.eclipse.chemclipse.xxd.classification.model.ClassificationDictionary;
 import org.eclipse.chemclipse.xxd.classification.model.ClassificationRule;
 import org.eclipse.chemclipse.xxd.classification.preferences.PreferenceSupplier;
